+++ components/Modal/ModalContainer.tsx
@@ -8,7 +8,7 @@ import useModalStore, { ModalName } from '@stores/useModalStore';
 import Portal from '@components/Portal';
 
 import styles from './ModalContainer.module.scss';
-import ReadPost from './ReadPost/ReadPost';
+import ReadPost from '../../app/map/_components/ReadPostModal';
 import WritePost from './WritePost/WritePost';
 
 // key 타입은 ModalName의 값, 값은 JSX.Element
