--- components/Modal/ReadPost/Body/ReadPostBody.module.scss
@@ -1,9 +1,0 @@
-@import '/styles/global-styles/media-breakpoint';
-
-@include tablet {
-  .main {
-    display: flex;
-    gap: 2.4rem;
-    overflow: hidden;
-  }
-}
