+++ components/ReadPost/Image/ImageContainer.tsx
@@ -7,7 +7,7 @@ import previousButton from '@/public/icons/previous-button.svg';
 import { Post } from '@/types/Post';
 
 import styles from './ImageContainer.module.scss';
-import useImageControl from '../../hooks/useImageControl';
+import useImageControl from '../hooks/useImageControl';
 
 interface ImageContainerProps {
   contentData: Post;
